adding termination-nodes feature

+++ test/helm/chart-test.sh
@@ -49,9 +49,20 @@ readonly CT_CONTAINER_NAME="ct"
 readonly CT_EXEC="docker exec --interactive $CT_CONTAINER_NAME"
 
 # AEMM
+readonly AEMM_HELM_REPO="$REPO_PATH/helm/amazon-ec2-metadata-mock"
 DOCKER_IMAGE_TO_LOAD="amazon-ec2-metadata-mock:test-latest"
 AEMM_DOCKER_IMAGE_INPUT=""
 DOCKER_ARGS=" --build-arg GOPROXY=direct "
+
+# Termination-Node test
+export CLUSTER_NAME
+export AEMM_HELM_REPO
+SCRIPTPATH="$(
+  cd "$(dirname "$0")"
+  pwd -P
+)"
+TEST_FILES=$(find $SCRIPTPATH/termination-test ! -name '*.yaml' -type f)
+export KUBECONFIG="$TMP_DIR/kubeconfig"
 
 readonly HELP=$(cat << 'EOM'
 Test Helm charts E2E by linting and/or installing helm charts in a provisioned environment. Only changed charts are tested.
@@ -71,6 +82,7 @@ Options:
   -g     AEMM image to use to test values.yaml file(s) with overridden image. See helm/amazon-ec2-metadata-mock/ci/custom-image-values.yaml
   -l     test charts for linting only (helm lint, version checking, YAML validation, maintainer validation)
   -i     test charts with installation only i.e. skip linting (deploys and runs helm test on charts for each *-values.yaml file in helm/<chart>/ci dir)
+  -t     test chart --termination-nodes functionality only
   -p     preserve the provisioned environment after test runs
   -r     reuse kind cluster and docker chart-testing container previously provisioned by this tool
   -d     debug, enables set -x, printing primary commands before executing
@@ -83,6 +95,7 @@ INSTALL_ONLY=false
 DEBUG=false
 PRESERVE=false
 REUSE_ENV=false
+TERMINATION_ONLY=false
 
 export TERM="xterm"
 RED=$(tput setaf 1)
@@ -113,9 +126,11 @@ create_kind_cluster() {
     c_echo "Creating kind Kubernetes cluster with kubeconfig in $KUBECONFIG_TMP_PATH"
     kind create cluster --name $CLUSTER_NAME --config $CLUSTER_CONFIG --image "kindest/node:$KIND_IMAGE" --kubeconfig $KUBECONFIG_TMP_PATH --wait 60s
 
-    c_echo "Copying kubeconfig to container..."
-    $CT_EXEC mkdir -p /root/.kube
-    docker cp $KUBECONFIG_TMP_PATH ct:/root/.kube/config
+    if [ $TERMINATION_ONLY == false ]; then
+      c_echo "Copying kubeconfig to container..."
+      $CT_EXEC mkdir -p /root/.kube
+      docker cp $KUBECONFIG_TMP_PATH ct:/root/.kube/config
+    fi
 
     c_echo "👍 Cluster ready!\n"
 }
@@ -142,7 +157,14 @@ handle_errors_and_cleanup() {
     echo
     MSG_PREFIX="-> "
     if [ $PRESERVE == true ]; then
-        c_echo "The test environment is preserved. Reuse with the '-r' option.\n${MSG_PREFIX}List Docker container: docker ps --filter \"name=^ct$\""
+        c_echo "The test environment is preserved. Reuse with the '-r' option.\n${MSG_PREFIX} List Docker container: docker ps --filter \"name=^ct$\""
+
+        c_echo "======================================================================================================"
+        c_echo "To poke around your environment manually:"
+        c_echo "export KUBECONFIG=$TMP_DIR/kubeconfig"
+        c_echo "export PATH=$TMP_DIR:\$PATH"
+        c_echo "kubectl get pods -A"
+        c_echo "======================================================================================================"
 
         if [[ $LINT_ONLY == true ]]; then
             c_echo "Cleanup commands:\n  * docker kill ct > /dev/null 2>&1" $MSG_PREFIX
@@ -154,8 +176,10 @@ handle_errors_and_cleanup() {
         fi
     else
         c_echo "Cleaning up resources..."
-        c_echo "Deleting ct container..." $MSG_PREFIX
-        docker kill ct > /dev/null 2>&1
+        if [ $TERMINATION_ONLY == false ]; then
+          c_echo "Deleting ct container..." $MSG_PREFIX
+          docker kill ct > /dev/null 2>&1
+        fi
 
          if [[ $LINT_ONLY == false ]]; then
             c_echo "Deleting kind cluster $CLUSTER_NAME..." $MSG_PREFIX
@@ -189,7 +213,11 @@ test_charts() {
         install_and_test_charts
     fi
 
-    if [ $LINT_ONLY == false ] && [ $INSTALL_ONLY == false ]; then
+    if [ $TERMINATION_ONLY == true ]; then
+        test_termination_nodes
+    fi
+
+    if [ $LINT_ONLY == false ] && [ $INSTALL_ONLY == false ] && [ $TERMINATION_ONLY == false ]; then
         lint_and_validate_charts
         install_and_test_charts
     fi
@@ -249,7 +277,7 @@ c_echo() {
 }
 
 process_args() {
-    while getopts "hdlprik:c:g:" opt; do
+    while getopts "hdlpritk:c:g:" opt; do
         case ${opt} in
             h )
               echo -e "$HELP" 1>&2
@@ -260,6 +288,7 @@ process_args() {
               ;;
             l )
               LINT_ONLY=true
+              c_echo "Running lint tests for Helm charts"
               ;;
             p )
               PRESERVE=true
@@ -269,6 +298,10 @@ process_args() {
               ;;
             i )
               INSTALL_ONLY=true
+              c_echo "Running E2E tests for Helm charts using the AEMM Docker image specified in values.yaml"
+              ;;
+            t )
+              TERMINATION_ONLY=true
               ;;
             k )
               OPTARG="K8s_$(echo $OPTARG | sed 's/\./\_/g')"
@@ -314,20 +347,26 @@ get_chart_test_config() {
     echo "$config"
 }
 
+test_termination_nodes() {
+    if [[ $REUSE_ENV == false ]]; then
+        mkdir -p $TMP_DIR
+        install_kind
+        create_kind_cluster
+    fi
+
+    build_and_load_image
+    for test_file in $TEST_FILES; do
+      $test_file
+    done
+}
+
 main() {
     process_args "$@"
 
     trap 'handle_errors_and_cleanup $? $BASH_COMMAND' EXIT
+    c_echo "Using:\n${BOLD}  * kind version=$KIND_VERSION\n  * Kubernetes version=$KIND_IMAGE\n  * helm/chart-testing version=$CT_TAG\n  * lint only=$LINT_ONLY\n  * install only=$INSTALL_ONLY\n  * termination only=$TERMINATION_ONLY\n  * preserve test env=$PRESERVE\n  * reuse=$REUSE_ENV\n  * debug=$DEBUG\n"
 
     chart_config=$(get_chart_test_config)
-
-    if [ $LINT_ONLY == true ]; then
-        c_echo "Running lint tests for Helm charts"
-        c_echo "Using:\n${BOLD}  * helm/chart-testing version=$CT_TAG\n  * lint only=$LINT_ONLY\n  * preserve test env=$PRESERVE\n  * reuse=$REUSE_ENV\n  * debug=$DEBUG\n"
-    else
-        c_echo "Running E2E tests for Helm charts using the AEMM Docker image specified in values.yaml"
-        c_echo "Using:\n${BOLD}  * kind version=$KIND_VERSION\n  * Kubernetes version=$KIND_IMAGE\n  * helm/chart-testing version=$CT_TAG\n  * lint only=$LINT_ONLY\n  * install only=$INSTALL_ONLY\n  * preserve test env=$PRESERVE\n  * reuse=$REUSE_ENV\n  * debug=$DEBUG\n"
-    fi
     echo -e "${MAGENTA}  From $CT_CONFIG:${BOLD}$chart_config"
     echo "${RESET_FMT}"
